examples/requirements.txt
anthropic>=0.49.0
mistralai>=1.5.0
python-dotenv>=1.0.1
